requirement.txt
einops==0.8.0
ftfy==6.1.1
h5py==3.12.1
huggingface-hub==0.36.0
idna==3.7
imageio==2.34.2
matplotlib==3.10.1
monai==1.3.1
nibabel==5.2.1
numpy==1.26.0
opencv-python==4.10.0.82
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
tensorboard==2.16.2
tensorboardX==2.6.2.2
timm==1.0.22
tokenizers==0.22.1
torchaudio==2.3.1
torchvision==0.18.1
tqdm==4.66.4
transformers==4.57.3
triton==2.3.1
wandb==0.19.8